age_core::format::read

Function legacy_age_stanza

pub fn legacy_age_stanza(input: &[u8]) -> IResult<&[u8], AgeStanza<'_>>

Reads a age stanza, allowing the legacy encoding of an body.

From the age spec:

Each recipient stanza starts with a line beginning with -> and its type name,
followed by zero or more SP-separated arguments. The type name and the arguments
are arbitrary strings. Unknown recipient types are ignored. The rest of the
recipient stanza is a body of canonical base64 from RFC 4648 without padding
wrapped at exactly 64 columns.

The spec was originally unclear about how to encode a stanza body. Both age and rage implemented the encoding in a way such that a stanza with a body of length of 0 mod 64 was indistinguishable from an incomplete stanza. The spec now requires a stanza body to always be terminated with a short line (empty if necessary). This API exists to handle files that include the legacy encoding. The only known generator of 0 mod 64 bodies is grease_the_joint, so this should only affect age files encrypted with beta versions of the age or rage crates.