age_core::primitives

Function bech32_encode_to_fmt

pub fn bech32_encode_to_fmt(f: &mut impl Write, hrp: Hrp, data: &[u8]) -> Result

Encodes data to a format writer as a Bech32-encoded string with the given HRP.

This implements Bech32 as defined in BIP-173, except it does not enforce the length restriction, allowing for a reduction in error-correcting properties.