Crate age_core

This crate contains common structs and functions used across the age crates.

You are probably looking for the age crate itself. You should only need to directly depend on this crate if you are implementing a custom recipient type.

Re-exports

pub use secrecy;

Modules

format

Core types and encoding operations used by the age file format.

io

Common helpers for performing I/O.

primitives

Primitive cryptographic operations used across various age components.